add geotiff reader into codebase
